requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==0.8.0
openai==1.35.3
python-dotenv==1.0.1
pydantic==2.8.2
